+++ R/globals.R
@@ -1,7 +1,7 @@
-utils::globalVariables(c("RMSE", "adjR", "control", "controlID",
+utils::globalVariables(c("RMSE", "adjR", "choice", "control", "controlID",
                          "control_coefs", "estimate", "estimate_FE", "feols",
-                         "fixed", "included", "index", "is_fe", "label", "lower",
-                         "observed", "outside", "p", "ribbon.group", "se",
-                         "se_type", "sig", "sig.level", "statistic", "term",
-                         "upper", "value", "."))
+                         "fixed", "included", "index", "is_fe", "is_residual",
+                         "label", "lower", "observed", "outside", "p", "percent",
+                         "ribbon.group", "se", "se_type", "sig", "sig.level",
+                         "statistic", "term", "upper", "value", "."))